requirements.txt
apispec==4.3.0
beautifulsoup4==4.9.3
Flask==1.1.2
flask-apispec==0.11.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
psycopg2-binary==2.8.6
requests==2.25.1
SQLAlchemy==1.3.23
